Issue-256: Moved FragmentViewBindingDelegate.kt file to core module
