+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/PaginatedScreenFragment.kt
@@ -12,9 +12,9 @@ import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
 import com.kirchhoff.movies.core.ui.recyclerview.Paginator
 import com.kirchhoff.movies.core.ui.recyclerview.decorations.GridMarginItemDecoration
+import com.kirchhoff.movies.core.ui.utils.viewBinding
 import com.kirchhoff.movies.data.ui.core.UIPaginated
 import com.kirchhoff.movies.databinding.FragmentPaginatedBinding
-import com.kirchhoff.movies.utils.viewBinding
 
 abstract class PaginatedScreenFragment<Data, T : UIPaginated<Data>> : BaseFragment(R.layout.fragment_paginated) {
