+++ core/build.gradle
@@ -1,6 +1,10 @@
 apply from: '../buildsystem/android-module.gradle'
 
 dependencies {
+    implementation libraries.lifecycle
+    implementation libraries.liveData
+    implementation libraries.viewModel
+
     implementation libraries.glide
     annotationProcessor libraries.glideCompiler
 }
